requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
_py-xgboost-mutex=2.0=cpu_0
blas=1.0=mkl
bottleneck=1.4.2=py313h2cb717b_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
colorama=0.4.6=py313haa95532_0
expat=2.7.1=h8ddb27b_0
fire=0.7.0=pypi_0
icc_rt=2022.1.0=h6049295_2
imbalanced-learn=0.13.0=py313haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
joblib=1.4.2=py313haa95532_0
libffi=3.4.4=hd77b12b_1
libmpdec=4.0.0=h827c3e9_0
libxgboost=2.1.1=h585ebfc_0
lightgbm=4.6.0=py313h5da7b33_0
loguru=0.7.2=py313haa95532_1
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py313h827c3e9_2
mkl_fft=1.3.11=py313h827c3e9_0
mkl_random=1.2.8=py313hce38976_0
numexpr=2.10.1=py313h4cd664f_0
numpy=2.2.5=py313ha127e54_0
numpy-base=2.2.5=py313haaa8b43_0
openssl=3.0.16=h3f729d1_0
pandas=2.2.3=py313h5da7b33_0
pip=25.1=pyhc872135_2
py-xgboost=2.1.1=py313haa95532_0
python=3.13.2=hadb2040_100_cp313
python-dateutil=2.9.0post0=py313haa95532_2
python-tzdata=2025.2=pyhd3eb1b0_0
python_abi=3.13=0_cp313
pytz=2024.1=py313haa95532_0
scikit-learn=1.6.1=py313h585ebfc_0
scipy=1.15.3=py313hde77213_0
setuptools=72.1.0=py313haa95532_0
six=1.17.0=py313haa95532_0
sklearn-compat=0.1.3=py313haa95532_0
sqlite=3.45.3=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
termcolor=3.1.0=pypi_0
threadpoolctl=3.5.0=py313h4442805_0
tk=8.6.14=h0416ee5_0
tzdata=2025b=h04d1e81_0
vc=14.42=haa95532_5
vs2015_runtime=14.42.34433=hbfb602d_5
wheel=0.45.1=py313haa95532_0
win32_setctime=1.1.0=pyhd3eb1b0_0
xgboost=2.1.1=py313haa95532_0
xz=5.6.4=h4754444_1
zlib=1.2.13=h8cc25b3_1
